player interactor interface and impl were removed, PlayerService interface was moved to domain layer
